Added Lever. Fixed SteamVR menu crash.
We reinitialize the hand every time the script is enabled now to account
for steamvr deactivating it.
There was a order of operations bug with the rotator and the deltamagic.
NVRLever was added, which launches a rocketship in the air when
completed.
Increased the drag of the door to make it wobble less.