Added Editor Preview Player and Playspace calls
In the editor we now have a representation of your current playspace bounds as well as how tall a 6ft (or 1.8m) player would be. This is to give you an idea of scale and orientation in the editor without popping into VR. We also have Integration.GetPlayspace() which will return a Vector3 of your placespace bounds. For SteamVR, y is always 1 on successful return. You can also access this with NVRPlayer.PlayspaceSize.