Moved button state to getters with cached values
Lazy loading the button states reduces the number of states we read each
update and only reads them when accessed individually. Also memoizes the
result for the remainder of the frame. I was seeing NVRHand.Update()
taking up to 2.5ms in the profiler, so this should help speed things up
a bit :)